fix problem with indexing array

+++ rfetch.h
@@ -4,7 +4,7 @@
 #ifndef RFETCH_H
 #define RFETCH_H
 
-std::string rget(const uint32_t &addr, const uint32_t regs[]);
+std::string rget(const size_t &addr, const uint32_t regs[]);
 
 std::vector<std::string> rfetch(const std::vector<std::string> &decoded, const uint32_t regs[32]);
